chore: 拷贝md文件，增加自动校验图片是否存在

+++ copy-images-md.mjs
@@ -1,6 +1,10 @@
 import { glob } from 'glob';
-import { copyFile, unlink } from 'fs/promises';
-import { basename } from 'path'
+import { copyFile, unlink, access, constants } from 'fs/promises';
+import { basename, dirname, join } from 'path'
+import { fileURLToPath } from 'url';
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(__filename);
 
 const mdFiles = await glob('./source/images/**/*.md');
 
@@ -10,11 +14,30 @@ let fileCount = 0;
 for (const file of mdFiles) {
     console.log(file);
     
+    // get the file dir name
+    const parentDir = dirname(file);
+
+    // check in the dir, has 5 JPEG file. eg: 1.jpg, 2.jpg, 3.jpg, 4.jpg, 5.jpg
+    let isSatisfy = true;
+    for (let i = 1; i <= 5; i++) {
+        const jpgFile = join(__dirname, parentDir, i + '.jpg');
+
+        try {
+            await access(jpgFile, constants.R_OK)
+        } catch {
+            isSatisfy = false;
+            break;
+        }
+    }
+
+    if (!isSatisfy) continue;
+
     // get the file name
     const fileName = basename(file);
 
     await copyFile(file, './source/_posts/' + fileName);
     await unlink(file);
+
     fileCount++;
 }
